+++ src/Main.h
@@ -38,8 +38,8 @@ extern "C" bool DllMain(HINSTANCE, int, void*) __attribute__((stdcall));
 #endif
 
 //These are intended for use with the SCAR/Simba plugin loaders, and nothing else.
-extern "C" long GetFunctionCount() __attribute__((stdcall));
-extern "C" long GetFunctionInfo(int, void*&, char*&) __attribute__((stdcall));
+extern "C" int GetFunctionCount() __attribute__((stdcall));
+extern "C" int GetFunctionInfo(int, void*&, char*&) __attribute__((stdcall));
 
 #endif	/* _MAIN_H */
